test/geometric_shapes: fix cone/cylinders assertions
When EPA is used on strictly convex shapes with very large intersections
(like two shapes perfectly blended with one another),
we can't use the default tolerance of 1e-6; it is way too low.
To converge to that precision, EPA would need a lot of iterations, a lot
of vertices and a lot of faces.
This is expected behavior.

The test cases were modified accordingly.