GJKSolver: bool `compute_penetration` skips EPA if needed
Sometimes we simply want to know whether two shapes are in collision or
not, or maybe we just want to know the distance between them only if the
two shapes are disjoint.
In these cases, we don't need to run the EPA algorithm.

This commit adds a boolean to all methods of GJKSolver. In the next
commit, this boolean will be set by collision and distance functions
depending on the contents of the collision/distance queries.