ColRequest: set `enable_contact` to true by default
Although this field is not used in hpp-fcl, the library behaves
as if `enable_contact` were true by default.
In the next commits, I will use this field to compute the contact
information only when `enable_contact` is true.
By setting this field to false, collision performance can be improved
if the only thing desired is a boolean collision result.