Core: only compute penetration info when needed
When using a `collide`-like function, the `enable_contact` and `security_margin`
fields of a `CollisionRequest` determine whether or not penetration information
is computed.
When using a `distance`-like function, the `enable_signed_distance` field of a
`DistanceRequest` determines whether to compute the penetration depth when
objects are intersecting.

Having these fields control the behavior of the underlying algos (GJK and EPA)
allows the user to control more finely the trade-off information vs. performance
they want from hpp-fcl.