Support func: remove useless `dirIsNormalized`
There is never the need to normalize the support direction when computing
a support point. This is because the argmax support function (the support
point) is invariant by multiplication by a positive scalar.
So the support in the direction dir or alpha * dir (with alpha > 0) will
always be identical, no matter the shape.