GJKSolver: remove useless `distance_upper_bound` field
GJKSolver has direct access to GJK's distance_upper_bound field, so
there is no need to store it in GJKSolver.
We don't remove max iterations and tolerance fields because they are
used in the constructor of GJK and EPA.
Of course we also don't remove things related to cached guess.