MinkowskiDiff: template support by `InflateSupport` boolean
GJK and EPA don't need to inflate the supports by the swept sphere radii
of the involved shapes at each their iterations.
This is a very nice mathematical property, which allows to run GJK/EPA on
the original shapes and simply correct the solution GJK/EPA found after
they have converged.

However, it is sometimes usefull to have access to the "true" support
function of a shape swept by a sphere.
Hence the template parameter.

The rule is simple: when interacting with GJK/EPA, this template parameter
should always be `false` (except for debugging purposes).
When interacting directly with a shape and using its support function,
the template parameter should always be `true`.