GJKSolver: remove useless `shapeIntersect` method
This method is not used in the code base and is harmful for developpment.
The reason is that this method is redundant with the `shapeDistance` method of `GJKSolver`
Also, this method template-specialized, which is a very bad idea.
The specialization of shape intersections **should not** be handled directly by the `GJKSolver`!
The specializations should be handled upstream (as it is already done), before calling methods from GJKSolver.

The GJKSolver should **only** be used to run GJK and EPA!

In general, when developping hpp-fcl, one should use `ShapeShapeCollide` or `ShapeShapeDistance`
from `internal/shape_shape_func.h`.
These functions are responsible for choosing whether to call a specialized function or call GJK+EPA.

******TLDR; DO NOT specialize the methods of GJKSolver!********
It is extremely handy in tests to be able to run GJK+EPA on any pair of shapes
and compare the results to specialized algorithms.
If the methods of GJKSolver were to be specialized, we lose this testing opportunity...