GJKSolver: return void instead of boolean
The GJKSolver estimates a distance between two shapes.
It's up to the user of GJKSolver to decide whether or not this estimate is
 low enough to be considered in collision.

In practice, functions like `ShapeShapeDistance` and `ShapeShapeCollide`
retrieve the distance estimate of GJKSolver and use the security margin
as well as a collision detection threshold to decide whether or not there is
a collision.