GJKSolver: restore previously deleted attributes
I previously deleted these attributes (woopsie).
It's actually important that they appear as attributes as most methods of
`GJKSolver` are const.
The idea is that only the API functions of hpp-fcl (like collide/distance or
ComputeCollision/Distance functors) are the only one functions which can set up
the GJKSolver. Then, any other internal functions (like ShapeShapeCollide or BVHShapeCollide etc.)
only work with `const GJKSolver *`. This way, we make sure that the settings of the GJKSolver are
not modified.