Swept-Sphere radius: clarify difference with "inflated" methods
- Certain primitives can be "inflated". This returns a pair of new primitive and a new Transform.
An inflated primitive is simply elongated/shrinked along its primary dimensions.
An inflated box is simply the original box which vertices have been elongated along the primary
axiis of the box.
Note that the inflation does not need to be positive. A primitive can be deflated ("inflated negatively").

- Swept-sphere radius is different: the shape is not modified. Swept-sphere radius is a property
of a shape which mainly softens its sharp corners (a box will become a box with rounded corners).
The swept-sphere radius cannot be negative.

To fully distinguish between these two features, I made sure any comment/template/variable referring
to swept-sphere radius did not have the word "inflation", as this word is reserved to talk about
the "inflated" methods of certain primitives.