EPA: fix normal in certain edge cases
Added a comment in EPA that explains it all.
In a future PR, I will remove the `NonConvex` status of EPA, as it useless.
Even when EPA is called "when it should not be" (i.e called on non-colliding shapes),
it will work just fine and return the correct result.
This is a very very cool property of EPA.