GJKSolver: fix doxygen to xml when empty template param
The doxygen_xml_parser.py script complains when an empty template
param is found (`typename = `). This happens for `GJKSolver::shapeDistance`
applied to TriangleP-TriangleP.
In any case, the `GJKSolver::shapeDistance` should not be template instanciated
for this pair of shapes as EPA does not work for TriangleP-TriangleP.
This fix actually makes sure the specialized TriangleP-TriangleP algo is used
instead.