ContactPatch: greatly simplify the struct
I did benchmarks, and it's completely useless to store the 2D polygon as
an Eigen matrix. It's much simpler to store it as a std::vector. It's therefore
super easy to deal with dynamic sizes internally, while remaining efficient.
I also reworked the support set functions API to make them hopefully clearer.

Note: the reason why we have the slightly convoluted `DEFAULT/INVERTED` directions
in the contact patch is to be able to express the support sets of shape1 and shape2
in the same frame. Otherwise it means that we have to do additional (and useless) transforms
which hurts performances.