CPatchSolver: rewrite Sutherland-Hodgman algo to handle all the edge cases
The SH algo is super easy to understand; however we have to handle the
edge cases carefully. The edge cases are mainly due to degenerate cases
where we have segments and points instead of non-empty convex polygons.
If not handled properly, these edge cases induce redundancies in the
final polygon of SH. For example, if SH is ran on two triangles which
have only one common point, SH will return an intersection polygon
composed of 3 identical points (the common point of the triangles).

We detect these cases by using a tolerance when computing determinants.
This allows us to remove redundant points.
Also, we deal with the special case of the segment-polygon.